Rename SortUtil.filterSort to filterSortableFields
